Rename file and small updates
